refactor: use id instead of tile struc/interface
The reason for this change is performance. The payload is ~5kb if the
entire structure is sent as json compared to 1.5kb if only the id is
sent.

Using initialised/hard coded arrays for lookup, we avoid creating and
storing structs at runtime. Instead the the TileList serve as
a read-only source of truth.